Support colorized as -G CLI option
This commit introduces CLI parsing along with support for the -G
(colorized) flag.

This also means the default behavior will no longer be to run the
program in colorized mode.

This commit is a little bit messier than I wanted it to be because
I decided to run a formatter on existing source files :(